ai-engine/requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
flask-cors==4.0.0
apify-client==1.5.0
